backend/requirements/develop.txt
-r base.txt
django-debug-toolbar>=2.2
six==1.13.0
django-silk>=4.0.1

coverage>=5.0.1
pep8==1.7.1
pylint>=2.4.4
pylint-quotes
pytest-factoryboy>=2.0.3
factory-boy>=2.12.0
django-sqlformatter>=1.0.3
mypy>=0.782
